pyproject.toml
[tool.poetry]
name = "meta"
version = "0.1"
description = "Advanced Telegram Bot API schema code generator"
authors = []
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
pydantic = "^1.9.1"
requests = "^2.28.1"
pyhumps = "^3.7.2"
jinja2 = "^3.1.2"
markdown = "^3.4.1"

[tool.poetry.dev-dependencies]
isort = "^5.12.0"
black = "^23.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
